requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pytz==2023.3

